Add compatibility for apklib & aar to be mixed together in an apk
